move ModuleAccountAddrs
